+++ common/src/main/java/net/xstopho/resource_cracker/item/ScytheItem.java
@@ -18,14 +18,14 @@ import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.block.CropBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.xstopho.resource_cracker.CrackerConstants;
+import net.xstopho.resource_cracker.config.CrackerConfig;
 
 import java.util.List;
-import java.util.UUID;
 
 public class ScytheItem extends SwordItem {
     public static final ResourceLocation BASE_ENTITY_REACH = ResourceLocation.fromNamespaceAndPath(CrackerConstants.MOD_ID, "base_entity_reach");
     public static final ResourceLocation BASE_BLOCK_REACH = ResourceLocation.fromNamespaceAndPath(CrackerConstants.MOD_ID, "base_block_reach");
-    private static final int radius = 1;
+    private static final int radius = CrackerConfig.SCYTHE_RADIUS.get();
 
     public ScytheItem(Tier tier, int attackDamage, float attackSpeed) {
         super(tier, new Properties()