+++ fabric/src/main/java/net/xstopho/resource_cracker/ResourceCrackerClient.java
@@ -4,10 +4,12 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.minecraft.client.renderer.RenderType;
 import net.xstopho.resource_cracker.registries.BlockRegistry;
+import net.xstopho.resource_cracker.registries.ItemModelRenderRegistry;
 
 public class ResourceCrackerClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         BlockRenderLayerMap.INSTANCE.putBlock(BlockRegistry.GARLIC_CROP.get(), RenderType.cutout());
+        ItemModelRenderRegistry.init();
     }
 }