+++ neoforge/src/main/java/net/xstopho/resource_cracker/ResourceCracker.java
@@ -7,19 +7,19 @@ import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
-import net.xstopho.resource_config_api.api.ConfigRegistry;
 import net.xstopho.resource_cracker.config.CrackerConfig;
 import net.xstopho.resource_cracker.modifier.LootModifier;
 import net.xstopho.resource_cracker.registries.BlockRegistry;
 import net.xstopho.resource_cracker.registries.CreativeTabRegistry;
 import net.xstopho.resource_cracker.registries.ItemModelRenderRegistry;
 import net.xstopho.resource_cracker.registries.ItemRegistry;
+import net.xstopho.resourceconfigapi.api.ConfigRegistry;
 
 @Mod(CrackerConstants.MOD_ID)
 public class ResourceCracker {
 
     public ResourceCracker() {
-        ConfigRegistry.register(CrackerConstants.MOD_ID, CrackerConfig.BUILDER);
+        ConfigRegistry.register(CrackerConstants.MOD_ID, CrackerConfig.BUILDER, true);
 
         BlockRegistry.init();
         ItemRegistry.init();