+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+#### v2.1.0
+- Scythe Radius config value now gets applied correctly
+- fixing Fabric Server crash because client code got executed on server
+- fixing visual bug for Crack Hammer and Chisel Tools
+  - durability gets applied when crafted
+  - config got synced between server and client to apply the server side durability
+
 #### v2.0.1
 - NeoForge:
   - adding missing Recipes for Mekanism and Applied Energetics 2